# docs/swagger.yaml
basePath: /
host: localhost:3000
info:
  contact: {}
  description: Backend API for Roha social platform.
  title: Roha Social API
  version: "1.0"
paths:
  /comments/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a comment. Only the commenter or the post owner can delete
        the comment.
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "Comment deleted successfully"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: '{"error": "Invalid comment ID"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: '{"error": "unauthorized to delete this comment"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: '{"error": "comment not found"}'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a comment
      tags:
      - comments
  /feed:
    get:
      description: Fetch personalized feed for authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: '{"error": "Unauthorized"}'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user feed
      tags:
      - posts
  /login:
    post:
      consumes:
      - application/json
      description: Login with email and password
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{"token": "JWT_TOKEN"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: '{"error": "Could not read request body"}'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - users
  /posts:
    get:
      description: Fetch all public posts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get all posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Create a post with media URL and caption
      parameters:
      - description: Post data
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error": "Invalid request body"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: '{"error": "Unauthorized"}'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new post
      tags:
      - posts
  /posts/{id}:
    delete:
      description: Delete a post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "Post deleted successfully"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: '{"error": "Invalid post ID"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: '{"error": "Unauthorized"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: '{"error": "Forbidden"}'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a post
      tags:
      - posts
    get:
      description: Fetch a post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error": "Invalid post ID"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: '{"error": "post not found"}'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a post
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Update post caption by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error": "Invalid post ID"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: '{"error": "Unauthorized"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: '{"error": "Forbidden"}'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a post
      tags:
      - posts
  /posts/{id}/comments:
    get:
      description: Fetch all comments belonging to a specific post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error": "Invalid post ID"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: '{"error": "post not found"}'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get comments for a post
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Add a comment to a specific post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment data
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error": "Invalid post ID"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: '{"error": "Unauthorized"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: '{"error": "post not found"}'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a comment
      tags:
      - comments
  /posts/{id}/like:
    delete:
      consumes:
      - application/json
      description: Removes a like relation between the authenticated user and a specific
        post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "Post unliked successfully"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: '{"error": "Invalid post ID"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: '{"error": "Unauthorized"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: '{"error": "post not found"}'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unlike a post
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Adds a like relation between the authenticated user and a specific
        post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "Post liked successfully"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: '{"error": "Invalid post ID"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: '{"error": "Unauthorized"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: '{"error": "post not found"}'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Like a post
      tags:
      - posts
  /signup:
    post:
      consumes:
      - application/json
      description: Register a new user with username, email, and password
      parameters:
      - description: User data
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "User created successfully"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: '{"error": "Could not read request body"}'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - users
  /users/{id}:
    get:
      description: Fetch user profile by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error": "Invalid user ID"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: '{"error": "User not found"}'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - users
  /users/{id}/follow:
    post:
      description: Follow another user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "Successfully followed user"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: '{"error": "Invalid user ID"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: '{"error": "Unauthorized"}'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Follow a user
      tags:
      - users
  /users/{id}/followers:
    get:
      description: Fetch followers of a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error": "Invalid user ID"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: '{"error": "User not found"}'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user followers
      tags:
      - users
  /users/{id}/following:
    get:
      description: Fetch users followed by a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error": "Invalid user ID"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: '{"error": "User not found"}'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user following
      tags:
      - users
  /users/{id}/unfollow:
    post:
      description: Unfollow another user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "Successfully unfollowed user"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: '{"error": "Invalid user ID"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: '{"error": "Unauthorized"}'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unfollow a user
      tags:
      - users
  /validate:
    get:
      description: Validate if the user's JWT token is valid and active
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: '{"error": "Unauthorized"}'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Validate JWT Token
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"